improve j1939 decoder
